requirements.txt
Flask==1.1.1
Flask-SocketIO==4.3.0
gunicorn==20.0.4
json5==0.8.5
Pillow==7.1.0
python-socketio==4.5.1
google-cloud-datastore==2.0.1
google-cloud-storage==1.33.0